rename exec file -> AppendEventInfo.cc
