+++ xPhoton/bin/py.AppendEventInfo.sample.py
@@ -4,7 +4,7 @@ fileList=[
 
 summaryfile='../data/mcInformation/summary_sigMC_amcatnlo.json'
 import os
-from xPhoton.xPhoton.AppendXSWeightToxPhtonTools import GetXS, nodir
+from xPhoton.xPhoton.AppendEventInfo import GetXS, nodir
 if __name__ == '__main__':
     newdir='updated'
     if not os.path.exists(newdir) or not os.path.isdir(newdir):
