+++ xPhoton/makefile
@@ -51,7 +51,8 @@ dev/exe.test.xPhotonRunner: obj/xPhotonRunner.o $(objfile_src) $(objfile_dev)
     obj/ExternalFilesMgr.o \
     obj/untuplizer.o \
     obj/usefulFuncs.o \
-	obj/BTagCalibrationStandalone.o
+	obj/BTagCalibrationStandalone.o \
+	obj/BTaggingMgr.o
 	@echo "--- Successfully compiled " $@ " ---"
 ###### declare test end ###########
 #obj/PreselectionMgr.o \
@@ -68,8 +69,8 @@ bin/exe.RooFit_ParFit_Pythia: obj/RooFit_ParFit_Pythia.o $(objfile_src)
 	$(Compiler) -o $@ $< $(objfile_src)
 bin/exe.xPhotonRunner: obj/xPhotonRunner.o $(objfile_src)
 	$(Compiler) -o $@ $< $(objfile_src)
-bin/exe.AppendXSWeightToxPhoton: obj/AppendXSWeightToxPhoton.o
-	$(Compiler) -o $@ $<
+bin/exe.AppendEventInfo: obj/AppendEventInfo.o $(objfile_src)
+	$(Compiler) -o $@ $< $(objfile_src)
 
 ################ Main Targets Ends #########################3
